Fix collections without doi_url
